Participation in the EW Project Challenge 2020
I am honored to have received a Certificate of Participation for my involvement in the EW Project Challenge 2020. My project, "Automated voice control car using esp32," was recognized for its innovative contribution to the field. The certificate is supported by esteemed institutions such as the NASSCOM Foundation, IEEE, Google Developers, TECHGIG, and the IEEE Computer Society, which makes this recognition especially significant. Issued on the 28th of June, 2020, by Ganesh Khomane, the Director of ElectronicWings, this acknowledgment serves as a valuable addition to my professional portfolio.